BackEnd/FypBackEndProject/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.2.6
django-admin-interface==0.17.1
django-ckeditor==6.1.0
django-colorfield==0.4.3
django-cors-headers==3.8.0
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-pandas==0.6.4
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
joblib==1.1.0
numpy==1.21.2
pandas==1.3.3
phonenumbers==8.12.35
Pillow==8.3.2
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
razorpay==1.2.0
requests==2.26.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sqlparse==0.4.1
threadpoolctl==3.0.0
urllib3==1.26.6
waitress==2.0.0
whitenoise==5.3.0
